feat: redirect to consult page on login and signup
